Move URLs to their own file
