+++ frontend/src/metabase-lib/queries/drills/dashboard-click-drill.js
@@ -8,7 +8,8 @@ import {
   getDataFromClicked,
   getTargetForQueryParams,
 } from "metabase-lib/parameters/utils/click-behavior";
-import Question, * as QUESTION from "metabase-lib/Question";
+import Question from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 
 export function getDashboardDrillType(clicked) {
   const clickBehavior = getClickBehavior(clicked);
@@ -110,10 +111,8 @@ export function getDashboardDrillQuestionUrl(question, clicked) {
   });
 
   return targetQuestion.isStructured()
-    ? QUESTION.getUrlWithParameters(targetQuestion, parameters, queryParams)
-    : `${QUESTION.getUrl(targetQuestion)}?${querystring.stringify(
-        queryParams,
-      )}`;
+    ? ML_Urls.getUrlWithParameters(targetQuestion, parameters, queryParams)
+    : `${ML_Urls.getUrl(targetQuestion)}?${querystring.stringify(queryParams)}`;
 }
 
 function getClickBehavior(clicked) {
