+++ frontend/src/metabase/components/MetadataInfo/TableInfo/ConnectedTables.tsx
@@ -2,7 +2,7 @@ import PropTypes from "prop-types";
 import { t } from "ttag";
 
 import Table from "metabase-lib/metadata/Table";
-import * as QUESTION from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 
 import { Label, LabelContainer, Container } from "../MetadataInfo.styled";
 import {
@@ -59,7 +59,7 @@ function ConnectedTableButton({
 
 function ConnectedTableLink({ table }: { table: Table }) {
   return (
-    <LabelLink to={QUESTION.getUrl(table.newQuestion())}>
+    <LabelLink to={ML_Urls.getUrl(table.newQuestion())}>
       <InteractiveTableLabel table={table} />
     </LabelLink>
   );
