+++ frontend/src/metabase/containers/AdHocQuestionLoader.unit.spec.js
@@ -5,7 +5,8 @@ import { delay } from "metabase/lib/promise";
 // import the un-connected component so we can test its internal logic sans
 // redux
 import { AdHocQuestionLoader } from "metabase/containers/AdHocQuestionLoader";
-import Question, * as QUESTION from "metabase-lib/Question";
+import Question from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 
 describe("AdHocQuestionLoader", () => {
   let loadQuestionSpy, loadMetadataSpy, mockChild;
@@ -22,7 +23,7 @@ describe("AdHocQuestionLoader", () => {
 
   it("should load a question given a questionHash", async () => {
     const q = Question.create({ databaseId: 1, tableId: 2 });
-    const questionHash = QUESTION.getUrl(q).match(/(#.*)/)[1];
+    const questionHash = ML_Urls.getUrl(q).match(/(#.*)/)[1];
 
     render(
       <AdHocQuestionLoader
