+++ frontend/src/metabase/models/components/ModelDetailPage/ModelInfoSidePanel/ModelInfoSidePanel.tsx
@@ -2,7 +2,7 @@ import { t } from "ttag";
 
 import type { Card } from "metabase-types/api";
 import type Question from "metabase-lib/Question";
-import * as QUESTION from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 import type Table from "metabase-lib/metadata/Table";
 
 import ModelRelationships from "./ModelRelationships";
@@ -58,7 +58,7 @@ function ModelInfoSidePanel({ model, mainTable, onChangeDescription }: Props) {
         <ModelInfoSection>
           <ModelInfoTitle>{t`Backing table`}</ModelInfoTitle>
           <ModelInfoLink
-            to={QUESTION.getUrl(mainTable.newQuestion(), { clean: false })}
+            to={ML_Urls.getUrl(mainTable.newQuestion(), { clean: false })}
             aria-label={t`Backing table`}
           >
             {mainTable.displayName()}
