+++ frontend/src/metabase/models/components/ModelDetailPage/ModelInfoSidePanel/ModelRelationships.tsx
@@ -5,7 +5,7 @@ import _ from "underscore";
 import Icon from "metabase/components/Icon";
 
 import type Question from "metabase-lib/Question";
-import * as QUESTION from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 import type Table from "metabase-lib/metadata/Table";
 
 import { ModelInfoTitle, ModelInfoSection } from "./ModelInfoSidePanel.styled";
@@ -37,7 +37,7 @@ function ModelRelationships({ model, mainTable }: Props) {
         {relatedTables.map(table => (
           <li key={table.id}>
             <ListItemLink
-              to={QUESTION.getUrl(table.newQuestion())}
+              to={ML_Urls.getUrl(table.newQuestion())}
               aria-label={table.displayName()}
             >
               <Icon name="table" />
