+++ frontend/src/metabase/models/components/ModelDetailPage/ModelUsageDetails/ModelUsageDetails.tsx
@@ -11,7 +11,7 @@ import Questions, {
 
 import type { State } from "metabase-types/store";
 import type Question from "metabase-lib/Question";
-import * as QUESTION from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 
 import {
   EmptyStateContainer,
@@ -41,7 +41,7 @@ function ModelUsageDetails({ model, questions, hasNewQuestionLink }: Props) {
           <EmptyStateActionContainer>
             <Button
               as={Link}
-              to={QUESTION.getUrl(model.composeDataset())}
+              to={ML_Urls.getUrl(model.composeDataset())}
               icon="add"
             >{t`Create a new question`}</Button>
           </EmptyStateActionContainer>
