+++ frontend/src/metabase/modes/components/drill/AutomaticInsightsDrill.tsx
@@ -46,7 +46,7 @@ export const AutomaticInsightsDrill = ({
       title: t`Compare to the rest`,
       section: "auto-popover",
       icon: "segment",
-      url: () => compareToRestDrillUrl({ question, clicked }),
+      url: () => compareToRestDrillUrl({ question, clicked }) as string,
     },
   ];
