+++ frontend/src/metabase/nav/components/QuestionLineage/QuestionLineage.tsx
@@ -1,7 +1,8 @@
 import { t } from "ttag";
 import Link from "metabase/core/components/Link/Link";
 import Badge from "metabase/components/Badge";
-import Question, * as QUESTION from "metabase-lib/Question";
+import Question from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 
 export interface QuestionLineageProps {
   icon?: string;
@@ -21,7 +22,7 @@ const QuestionLineage = ({
   return (
     <Badge icon={icon} isSingleLine>
       {t`Started from`}{" "}
-      <Link className="link" to={QUESTION.getUrl(originalQuestion)}>
+      <Link className="link" to={ML_Urls.getUrl(originalQuestion)}>
         {originalQuestion.displayName()}
       </Link>
     </Badge>
