+++ frontend/test/metabase-lib/lib/Question.unit.spec.js
@@ -25,7 +25,8 @@ import {
   SAMPLE_DB_ID,
 } from "metabase-types/api/mocks/presets";
 import { TYPE as SEMANTIC_TYPE } from "cljs/metabase.types";
-import Question, * as QUESTION from "metabase-lib/Question";
+import Question from "metabase-lib/Question";
+import * as ML_Urls from "metabase-lib/urls";
 import StructuredQuery from "metabase-lib/queries/StructuredQuery";
 import NativeQuery from "metabase-lib/queries/NativeQuery";
 
@@ -896,11 +897,11 @@ describe("Question", () => {
           assoc(orders_raw_card, "id", 1),
           metadata,
         );
-        expect(QUESTION.getUrl(question)).toBe("/question/1-raw-orders-data");
+        expect(ML_Urls.getUrl(question)).toBe("/question/1-raw-orders-data");
       });
       it("returns a URL with hash for an unsaved question", () => {
         const question = new Question(dissoc(orders_raw_card, "id"), metadata);
-        expect(QUESTION.getUrl(question)).toBe(adhocUrl);
+        expect(ML_Urls.getUrl(question)).toBe(adhocUrl);
       });
     });
 
@@ -910,7 +911,7 @@ describe("Question", () => {
         metadata,
       );
 
-      expect(QUESTION.getUrl(question)).toBe(adhocUrl);
+      expect(ML_Urls.getUrl(question)).toBe(adhocUrl);
     });
   });
 
@@ -1357,7 +1358,7 @@ describe("Question", () => {
         const parameters = [];
         const parameterValues = {};
 
-        const url = QUESTION.getUrlWithParameters(
+        const url = ML_Urls.getUrlWithParameters(
           question,
           parameters,
           parameterValues,
@@ -1371,7 +1372,7 @@ describe("Question", () => {
       });
 
       it("should return question URL with string MBQL filter added", () => {
-        const url = QUESTION.getUrlWithParameters(question, parameters, {
+        const url = ML_Urls.getUrlWithParameters(question, parameters, {
           1: "bar",
         });
 
@@ -1392,7 +1393,7 @@ describe("Question", () => {
       });
 
       it("should return question URL with number MBQL filter added", () => {
-        const url = QUESTION.getUrlWithParameters(question, parameters, {
+        const url = ML_Urls.getUrlWithParameters(question, parameters, {
           5: 123,
         });
 
@@ -1411,7 +1412,7 @@ describe("Question", () => {
       });
 
       it("should return question URL with date MBQL filter added", () => {
-        const url = QUESTION.getUrlWithParameters(question, parameters, {
+        const url = ML_Urls.getUrlWithParameters(question, parameters, {
           3: "2017-05",
         });
 
@@ -1431,7 +1432,7 @@ describe("Question", () => {
 
       it("should include objectId in a URL", () => {
         const OBJECT_ID = "5";
-        const url = QUESTION.getUrlWithParameters(
+        const url = ML_Urls.getUrlWithParameters(
           question,
           parameters,
           { 1: "bar" },
@@ -1450,7 +1451,7 @@ describe("Question", () => {
       const question = new Question(card);
 
       it("should return a card with attached parameters and parameter values as query params", () => {
-        const url = QUESTION.getUrlWithParameters(question, parameters, {
+        const url = ML_Urls.getUrlWithParameters(question, parameters, {
           1: "bar",
         });
 
@@ -1471,7 +1472,7 @@ describe("Question", () => {
       });
 
       it("should not include objectId in a URL", () => {
-        const url = QUESTION.getUrlWithParameters(
+        const url = ML_Urls.getUrlWithParameters(
           question,
           parameters,
           { 1: "bar" },
@@ -1521,7 +1522,7 @@ describe("Question", () => {
       const question = new Question(cardWithTextFilter, metadata);
 
       it("should return question URL when there are no parameters", () => {
-        const url = QUESTION.getUrlWithParameters(question, [], {});
+        const url = ML_Urls.getUrlWithParameters(question, [], {});
         expect(parseUrl(url)).toEqual({
           pathname: "/question/1",
           query: {},
@@ -1530,7 +1531,7 @@ describe("Question", () => {
       });
 
       it("should return question URL with query string parameter when there is a value for a parameter mapped to the question's variable", () => {
-        const url = QUESTION.getUrlWithParameters(
+        const url = ML_Urls.getUrlWithParameters(
           question,
           parametersForNativeQ,
           {
@@ -1547,7 +1548,7 @@ describe("Question", () => {
 
       it("should return question URL with query string parameter when there is a value for a parameter mapped to the question's field filter", () => {
         const question = new Question(cardWithFieldFilter, metadata);
-        const url = QUESTION.getUrlWithParameters(
+        const url = ML_Urls.getUrlWithParameters(
           question,
           parametersForNativeQ,
           {
@@ -1563,7 +1564,7 @@ describe("Question", () => {
       });
 
       it("should not include objectId in a URL", () => {
-        const url = QUESTION.getUrlWithParameters(
+        const url = ML_Urls.getUrlWithParameters(
           question,
           parametersForNativeQ,
           {
